pyworkspace/nnet/requirements.txt
torch
matplotlib
pyserial